// src/types/ColorState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { HslAdjust } from "./HslAdjust";
import type { HslBand } from "./HslBand";

export type ColorState = { vibrance: number, saturation: number, hsl: { [key in HslBand]?: HslAdjust }, bw: boolean, };