// src/types/EffectsState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type EffectsState = { clarity: number, dehaze: number, vignetteAmount: number, vignetteMidpoint: number, vignetteRoundness: number, vignetteFeather: number, grainAmount: number, grainSize: number, grainRoughness: number, };